convert cog.jsx to TypeScript
